requirements.txt
boto3
botocore
jmespath
python-dotenv
pytest
pytest-cov
pytest-mock
